Ajuste de correção de caracteres na bronze layer e ajustes nas DAGs.
