+++ air_flow/src/bronze_layer_full.py
@@ -2,6 +2,21 @@ import requests
 import json
 import os
 import time
+import unicodedata
+
+def normalize_text(text):
+    if not isinstance(text, str):
+        return text
+    try:
+        # Remove acentos/caracteres não-ASCII (defesa básica)
+        nfkd = unicodedata.normalize('NFKD', text)
+        return ''.join([c for c in nfkd if not unicodedata.combining(c)])
+    except Exception:
+        return text
+
+def normalize_record(record):
+    # Normaliza cada campo string do registro (dict)
+    return {k: normalize_text(v) if isinstance(v, str) else v for k, v in record.items()}
 
 def fetch_raw_data():
     try:
@@ -17,18 +32,20 @@ def fetch_raw_data():
             data = response.json()
             if not data:
                 break
+            # Normaliza cada registro recebido
+            data = [normalize_record(rec) for rec in data]
             all_breweries.extend(data)
             if len(data) < per_page:
-                break  # Chegou ao final
+                break
             page += 1
-            time.sleep(0.1)  # delay para evitar sobrecarga no servidor
+            time.sleep(0.1)
 
         os.makedirs('/opt/airflow/files/bronze/', exist_ok=True)
-        with open('/opt/airflow/files/bronze/breweries_raw.json', 'w') as f:
-            json.dump(all_breweries, f, indent=2)
+        with open('/opt/airflow/files/bronze/breweries_raw.json', 'w', encoding='utf-8') as f:
+            json.dump(all_breweries, f, indent=2, ensure_ascii=False)
 
         print(f"[BRONZE] Total de cervejarias baixadas: {len(all_breweries)}")
-        print("[BRONZE] Dados brutos foram salvos em /opt/airflow/files/bronze/breweries_raw.json")
+        print("[BRONZE] Dados brutos normalizados e salvos em /opt/airflow/files/bronze/breweries_raw.json")
     except Exception as e:
         print(f"[ERRO] Falha ao buscar dados brutos: {e}")
